requirements.txt
Django==3.1
psycopg2-binary==2.8.5
django-environ==0.4.5
djangorestframework==3.11.1
social-auth-app-django==4.0.0
drf-yasg==1.17.1
django-oauth-toolkit==1.3.2
djoser==2.0.5
